requirements.txt
Django==4.2.23
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
drf-spectacular==0.27.0
discord.py==2.3.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9 